pyproject.toml
[project]
name = "pdf-annotation-tool"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.24.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "types-PyYAML>=6.0",
]

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "pytest>=8.0.0",
    "types-PyYAML>=6.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_annotation_tool"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
exclude = ["tools/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
